bubasics: use config variable $(AR) to build libbubasics.a
Since the original IBM code import, libbubasics.a has been built with a
hardcoded 'ar' instead of the customary $(AR).

For consistency with the rest of the tree (and in preparation for a
future commit), convert this to use $(AR) instead.

While here, also change the options from 'r' to the more typical 'crv'.

Change-Id: If74e4b92188af9de7f504e3f588b7ee1070425fc
Reviewed-on: https://gerrit.openafs.org/14895
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>