bozo: Check for negative simple bnode parameter index
The simple (ez) bnodes have only one parameter, which is accessed as
index 0.  A check is made for values greater than 0, however the index
argument is a signed type, and a check for negative indices is missing.

Explicitly check for non-zero index values when retrieving simple bnode
parameters, and if not zero return BZDOM.

The other bnode types already check for specific (positive) index
values, so only the simple bnode type needs to be updated.

Change-Id: I8764af6c975431caf26c25a371cf6db10e6e7fec
Reviewed-on: https://gerrit.openafs.org/15043
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>