vos: Check start-of-dump magic in vos restore
When restoring with a regular file (either with -file or a shell redirect
'<'), vos restore verifies the file ends with the end-of-dump magic
number. This catches cases in which the specified file is obviously not
a dump file, for example the file is empty, truncated, or just not file
created by vos dump.

Currently, no checks are done when restoring from a pipe or fifo, since it
is not possible to check the end of the stream and reset the seek pointer
to before sending the data to the volume server.

Add a check to verify the dump begins with the dump header tag and the
start-of-dump magic number, even when the dump file is not seekable.

In the case the dump file is not seekable, send the start of dump
tag and the start-of-dump magic number directly in SendFile(), since
those values were already read in CheckDumpFile().

This guards against starting a restore from an empty or wrong file when
restoring from a pipeline or a named pipe.

Change-Id: I7aa508bf5f0e81da7d6602856b5242cb6313e9a8
Reviewed-on: https://gerrit.openafs.org/14711
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>